+++ split_fastq/HISTORY.md
@@ -1,6 +1,15 @@
 Release History
 ===============
 
+1.0.1
+-----
+
+**Features and Enhancements**
+
+- Using output prefix instead of output directory
+- Output interval file contains suffixes instead of numbers
+
+
 1.0.0
 -----